Read and reviewed
I read this over and it looks great.  I made a couple of changes, most of which are small.  The only somewhat substantive change is in the `Probability notation and Bayes' rule` section, where I updated the modeling and language to be more analogous to an RCT.  It's tough to get right, because readers may struggle to connect the abstract p(\theta | y) to as RCT (\theta is a vector of parameters, y is a matrix of observed data...); but if we tailor it too much to an RCT we lose generality.  Would love to hear y'all's thoughts.